# Improve proto.Buffer pool performance