--- a/mixer/pkg/protobuf/yaml/pool.go
+++ b/mixer/pkg/protobuf/yaml/pool.go
@@ -31,7 +31,6 @@ const (
 	steps      = 20
 
 	minSize = 1 << minBitSize
-	maxSize = 1 << (minBitSize + steps - 1)
 
 	calibrateCallsThreshold = 42000
 	maxPercentile           = 0.95