@istio-testing 2020-01-09
@panjf2000: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | 971268218a6243835ae8f619c5ed6ea7d3423162 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19989/lint_istio/5816) | `/test lint_istio`
unit-tests_istio | 971268218a6243835ae8f619c5ed6ea7d3423162 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19989/unit-tests_istio/6070) | `/test unit-tests_istio`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->