@istio-testing 2020-01-09
@panjf2000: The following test **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-istioio-k8s-tests_istio | 3fae5ef4fbdee2263d33e2923a0945f198f13203 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19989/integ-istioio-k8s-tests_istio/2908) | `/test integ-istioio-k8s-tests_istio`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->